관광지 읽기페이지 리뷰 리스트 띄우기

+++ src/main/java/com/example/board/model/review/Review.java
@@ -11,7 +11,7 @@ import lombok.Data;
 public class Review {
 	
 	private Long review_id;
-	private Long UC_SEQ; // 
+	private int UC_SEQ; // 
 	private String member_id;
 	private String nickname;
 	private String sort; // 관광지, 음식점, 호텔 중 택1
